Move WooSavedPackagesSelectionView to separate file
